+++ CodexLib/Const.cs
@@ -1,0 +1,15 @@
+﻿using Kingmaker.Enums;
+using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace CodexLib
+{
+    public class Const
+    {
+        public const ModifierDescriptor Intelligence = (ModifierDescriptor)5380;
+        public const ModifierDescriptor Charisma = (ModifierDescriptor)5381;
+    }
+}