+++ CodexLib/MasterPatch.cs
@@ -28,6 +28,10 @@ namespace CodexLib
 
             PatchSafe(typeof(Patch_ContextStatValue));
             PatchSafe(typeof(Patch_GetTargetProjectileFix));
+            //PatchSafe(typeof(Patch_RulebookEventBusPriority));
+
+            Helper.EnumCreateModifierDescriptor(Const.Intelligence, "Intelligence", "");
+            Helper.EnumCreateModifierDescriptor(Const.Charisma, "Charisma", "");
 
             harmony = null;
         }